+++ R/kernel_ram.R
@@ -35,7 +35,7 @@
 #' @references 
 #' Vihola, M. (2012). Robust adaptive Metropolis algorithm with coerced acceptance
 #' rate. Statistics and Computing, 22(5), 997–1008.
-#' \url{https://doi.org/10.1007/s11222-011-9269-5}
+#' \doi{10.1007/s11222-011-9269-5}
 #' @family kernels
 #' @examples 
 #' # Setting the acceptance rate to 30 % and deferring the updates until